Fixed a bug in input generation caused by overflows
Added package info

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/ranged_execution@41780 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/core/algorithm/rangedExecInput/InputGenerationWithRandomWalkAlgorithm.java
@@ -148,7 +148,7 @@ public class InputGenerationWithRandomWalkAlgorithm implements Algorithm {
     for (int i = 0; i < pComputedPath.size(); i++) {
       ARGPath path = pComputedPath.get(i);
       try {
-        List<Pair<CIdExpression, Integer>> inputs =
+        List<Pair<CIdExpression, Long>> inputs =
             utils.computeInputForRandomWalkPathNonIterative(
                 AbstractStates.extractStateByType(path.getLastState(), RandomWalkState.class),
                 path);
